pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tstkit"
version = "0.1.0"
description = "NEB calculation processing and KMC simulation toolkit"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",
    "ase>=3.22.0",
    "numba>=0.55.0",
    "matplotlib>=3.5.0",
    "pyyaml>=6.0",
]

[project.scripts]
tstkit = "tstkit.main:main"
#tstkit-readfreq = "tstkit.read_freq:main"
#tstkit-calcrate = "tstkit.calc_rate:main"

[tool.setuptools.packages.find]
where = ["src"]